Move coverage from DroidSansFallback to the Symbols font.
The new Noto Sans Symbols subsetted font adds all characters from
DroidSansFallback coverage that are already supported in the complete Noto
Sans Symbols font but are not available in Roboto.  This change opens up the
opportunity to subset DroidSansFallback to remove all characters already
supported in Roboto or the new Symbols subset, and remove the fallback font
from the extended font profile altogether.

Change-Id: I450eaaa5c40f72ddfa68104bfef72dd2edcee9ee