Update Noto Color Emoji to version 1.01 to include U+303D.
U+303D PART ALTERNATION MARK is the only character remaining in the black
and white AndroidEmoji font that is used. This makes it possible for
AndroidEmoji to be removed.

Additionally, the information in the name table is corrected and improved,
and the post table has been converted to version 3.0 to reduce font size.
Altogether, file size is reduced by 6KB.

Bug: 15777351
Change-Id: Ic868d9f56dd286c17df7371b46808a2cd23e4921