Expand symbols and Japanese coverage.
A slightly larger subset of Noto Sans Symbols is included on all
devices, to complete support for WGL4 and a few more frequently used
characters (the impact on font file size is 2kB).

A larger subset of Noto Sans Symbols is included on devices on the
extended font profile to also complete the support for the ARIB
character set, and as much as possible, the following Unicode blocks
(chosen based on frequency of characters on the web): Combining
Diacritical Marks for Symbols, Letterlike Symbols, Mathematical
Operators, Miscellaneous Technical, Control Pictures, Optical
Character Recognition, Geometric Shapes, Miscellaneous Symbols,
Supplemental Arrows-A, and Supplemental Mathematical Operators.

A subset of NotoSansJP is added for the devices on the extended font
profile, to match KLP character coverage and complete support for the
following Unicode blocks (as much as available in NotoSansJP): CJK
Symbols and Punctuation, Hiragana, Katakana, Katakana Phonetic
Extensions, Enclosed CJK Letters and Months, CJK Compatibility, and
CJK Unified Ideographs Extension B, C, and D.

The total impact on font file size is 296kB for the extended font
profile and 2kB on the non-extended one.

Bug: 4673396
Bug: 13680438
Bug: 16314977
Bug: 16856221
Bug: 17208391
Change-Id: I1f3fb87f3828151c6b439b63a73abc1eac591212