Remove default emoji characters from CJK fonts.
Twenty-six charactes that Unicode Technical Report #51 "Unicode
Emoji" defines as defaulting to emoji styles used to be displayed as
black and white ("text" style) before this. This patch removes those
characters from Noto CJK fonts, so they get displayed as color.

As a side effect of the subsetting process, some ideographic
variation sequences are also removed from some of the fonts, but this
doesn't cause a difference, as they are not presently supported in
Minikin.

The Korean font doesn't change in coverage, but seems to get some
size reduction being passed through fontTools, so it's included in
the patch.

The fonts are subsetted by the following Python script:

from nototools import subset

EMOJI = (
    [0x26BD, 0x26BE, 0x1F18E]
    + range(0x1F191, 0x1F19A+1)
    + [0x1F201, 0x1F21A, 0x1F22F]
    + range(0x1F232, 0x1F236+1)
    + [0x1F238, 0x1F239, 0x1F23A, 0x1F250, 0x1F251]
)

for variation in ['JP', 'KR', 'SC', 'TC']:
    font_file_name = 'NotoSans%s-Regular.otf' % variation
    subset.subset_font(font_file_name,
                       'no-emoji/'+font_file_name,
                       exclude=EMOJI)

Bug: 17244669
Change-Id: I48cc05b2bbcc5acd4540394dc5778cc63d454081