Update makefiles for new fonts.
- Tibetan Bold, Symbols-Regular-Subsetted2

Update Noto fonts from upstream.

This adds a new subset of the Symbols font.  In addition, it adds the
following new Noto fonts:
  NotoNastaliqUrdu-Regular.ttf
  NotoSansTibetan-Bold.ttf
  NotoSerifBengali-Bold.ttf
  NotoSerifBengali-Regular.ttf
  NotoSerifGujarati-Bold.ttf
  NotoSerifGujarati-Regular.ttf
  NotoSerifKannada-Bold.ttf
  NotoSerifKannada-Regular.ttf
  NotoSerifMalayalam-Bold.ttf
  NotoSerifMalayalam-Regular.ttf
  NotoSerifTamil-Bold.ttf
  NotoSerifTamil-Regular.ttf
  NotoSerifTelugu-Bold.ttf
  NotoSerifTelugu-Regular.ttf

These fonts were built using the upstream Makefile android target.  This
scales the Malayalam and Tamil UI fonts and compresses the others.
Then a tool was run upstream to build the Symbols subsets and to add
hyphens to Ethiopic and Armenian.

Bug: 27531263
Bug: 26862061
Change-Id: Iab359863f11c8afe886ed4b3f2aab76e3650ecb9