Update the subsetting script for Noto Sans Symbols
The symbols subsetting script had fallen out of sync with the actual
fonts it generated.

This CL adds the Japanese TV symbols that were already included in
the subset, and moves away from an explicitly-included default emoji
list to the one coming from nototools' unicode_data.

We also add all default-emoji characters that we are removing from
the symbols font to the second subset, so they can be accessed using
text variation sequences or setting the text language to request
non-emoji forms. Note that although text variation sequences are not
defined for all of those characters, they may get added in the near
future and even if they don't, Android has other ways of requesting
non-emoji forms.

Also update the subsetted fonts by running:
./subset_noto_sans_symbols.py NotoSansSymbols-Regular.ttf \
   NotoSansSymbols-Regular-Subsetted.ttf \
   NotoSansSymbols-Regular-Subsetted2.ttf

The first font is not changed except for time stamps, but the second
now includes various new characters.

Bug: 27531263
Bug: 32046261
Test: Dumped the fonts using ttx and diffed them
Change-Id: Iaf3f8b38feef05d48b9f8b8712f0b9f857109d0f