Lighten flag colors in Emoji font.
The flags have a shadow effect, the change lightens them a bit so this
should be better.  Because of the shadow effect the brightness will vary
across the image.

Tested the Russian flag using Gimp color picker.  Also did a side-by-side
comparison of all the flags, they all look a little bit lighter.

Updated version number in README.android.

Change-Id: Ic761858f2afe848b08370334e14cc23745cb35b0
Fixes: 37375651  The Russian flag emoji has an incorrect blue color
Test: visual inspection